rename ressource dir to resource
