+++ test/resource/dudes_profile_expected_output_on_sampledata.out
