+++ test/resource/dudesdb.npz
