+++ test/resource/idmapping_uniparc-head.tsv
