+++ test/resource/uniparc_filtered.fasta
