+++ test/resource/uniprot_sprot-head.fasta
