+++ test/test_DUDes.py
@@ -1,12 +1,12 @@
 from dudes.main import main
-from test.helper_funcs import RESSOURCE_DIR, DUDES_DIR, SAMPLEDATA_DIR, md5sum, set_directory
+from test.helper_funcs import RESOURCE_DIR, DUDES_DIR, SAMPLEDATA_DIR, md5sum, set_directory
 from unittest.mock import patch
 import sys
 
 
 def test_DUDes(tmp_path):
     produced_output = tmp_path / "dudes_profile_output"
-    expected_output = RESSOURCE_DIR / "dudes_profile_expected_output_on_sampledata.out"
+    expected_output = RESOURCE_DIR / "dudes_profile_expected_output_on_sampledata.out"
     args = ["dudes",
             "-s", str(SAMPLEDATA_DIR.relative_to(DUDES_DIR) / "hiseq_accuracy_k60.sam"),
             "-d", str(SAMPLEDATA_DIR.relative_to(DUDES_DIR) / "arc-bac_refseq-cg_201503.npz"),
